Changed the grid_type variable names to G.
  Following MOM6, the grid type variables are all set to simply "G".  This is a
valid exception to the proscription against single letter variables because
the grid type variables are almost always used as G% or as arguments, both of
which are easily tracked.  The advantage is that with having several references
to G on a single line being common, this keeps the code short and readable.
Also nullified the ice_dyn_CS pointer in the ice_type in the declaration, which
appears to eliminate some episodic segmentation faults with the gnu compiler.
All answers are bitwise identical.