Eliminated diagnostic arguments from ice_dynamics.
  Eliminated fx_ice, fy_ice, fx_cor, fy_cor from the call to ice_dynamics, since
these are strictly diagnostic arrays that are only used inside that subroutine.
Also eliminated the public declaration for the interfaces to ice_strength, sigI
and sigII, and changed the sizes of a number of arrays from the computational
domain size to the data domain size.  All answers are bitwise identical, but
the public interface to ice_dynamics is changed.