Use post_SIS_data for diagnostics in ice_dyn.
  Changed the call to send diagnostics to post_SIS_diag in the ice_dyn module,
which avoids copy-in/copy-out for calls operating on arrays with halos.  Also
changed the definitions of diag%is, etc., in SIS_diag_ctrl to allow for arrays
that do not start at 1.  This also requires calls to enable_SIS_averaging and
disable_SIS_averaging in teh ice model itself. All answers are bitwise
identical, and no interfaces are changed.  -RWH