Moved sig11, sig12, and sig22 into ice_dyn module.
  The stress tensors are only used in the ice dynamics module, and the spatial
extent of these variables will be specific to the spatial discretization of the
momentum equations, so they have been incorporated into ice_dyn.F90.  This has
also led to the addition of two new subroutines, ice_dyn_register_restarts and
ice_dyn_end, which need to be called from outside this module, and the removal
of the sig## arguments from ice_dynamics.  Answers are bitwise identical, but
there are substantial interface changes.  -RWH