Added G%CatIce, G%NkIce, and G%NkSnow.
  Added variables in the sea ice grid type to specify the number of sea ice
thickness categories and the number of vertical layers used to describe the
thermodynamic properties of the sea ice and snow.  This included adding the run
time parameters NCAT_ICE, NK_ICE, and NK_SNOW, to specify these values.  They
are checked for consistency with static memory.
  Also eliminated the unused functions dTdx and dTdy and added param_file as an
argument to set_ice_grid.  All answers are bitwise identical, although the
interface to set_ice_grid is changed. -RWH