Cosmetic clean-up of ice_model.F90.
  Made numerous cosmetic changes to ice_model.F90, which collectively resulted
in an ~10% shortening of the file. Merged 33 sets concentric do loops onto the
same lines. Changed the white space on 914 lines.  Added explicit colons in 9
instances where implied array syntax was used to set arrays, and in-lined the
function ext().  All answers are bitwise identical, and no interfaces are
changed. -RWH