Fixed add_diurnal_SW & made Ice%coszen a 2-d array.
  With add_diurnal_SW, the ocean had recieved an arteficial diurnal cycle of
shortwave radiation, but the sea ice had not.  This has now been changed so that
the two components receive consistent shortwave radiation.  In addition,
Ice%coszen had previously been declared with memory for every thickness
category, but of course the solar zenith angle is invariant across categories,
so this is now just a 2-d array.  Answers will change with add_diurnal_SW. -RWH