Combined update_ice_model_fast routines.
  Eliminated the interface to subroutine update_ice_model_fast_old, instead
merging it directly with updated_ice_model_fast.  This was how it already
appeared from the outside, so this is just streamlining the internal code.  Also
renamed some variables in this routine without the suffix "_new" as this was
no longer needed to differentiate from the arguments that are no longer used.
All answers and public interfaces are unchanged. -RWH