Moved ice_model_nml code into checking routine.
  Ice_model_nml no longer serves any purpose in specifying the parameters for
SIS2, however it is important to check for attempts to use it, so it still is
read just for verification against the parameters that are being set from
SIS_input.  Also removed some unused variables and regrouped the deallocate
statements in the same order as the allocates and added deallocates for some
arrays that were not being deallocated.  Apart from no longer using
ice_model_nml, external interfaces are unchanged and all answers are bitwise
identical.