Use Ice%area to initialize the iceberg cell areas.
  Replaced the nondimensional cell_area, with units of a fraction of the sphere
with Ice%area, with the more appropriate units of m2.  This leads to answer
changes at the level of roundoff, and the old and new interfaces can not be
mixed.  Also changed MOM_input from using N[IJ]PROC_IO to IO_LAYOUT for SIS2
and SIS2_icebergs.  New reference solutions were checked in for SIS2_icebergs.