Imported SIS_diag_mediator changes from MOM6.
  Added an axesType and the automatic application of land-masks to all variables
that are output via the diag_mananger, using the capabilities that have recently
been added to MOM6.  All answers are bitwise identical and answers unchanged,
though some output diagnostics now have masks applied, and the axes... elements
of the diag type are now axesTypes and not simple arrays of integers.