Added CGRID_ICE_DYNAMICS & C-grid elements to ice_type.
  Added c-grid ice and ocean velocity tomponents, along with the logical
Cgrid_dyn to the ice_type, and the new run-time parameter CGRID_ICE_DYNAMICS
that will eventually be used to select the discretization of the dynamic core.
Also started adding commented out code for use when the C-grid dynamics are in
use.  Some index cases were also corrected, and tmp1 and tmp2 were replaced with
hi_avg and hs_avg for greater clarity.  All answers are bitwise identical, and
no code needs to be updated due to interface extensions.