Added handling of various stress staggerings.
  Added code to allow for the run-time specification of the staggering of the
ocean velocities provided to the sea-ice and of the stress provided to the
ocean.  The new parameter ICE_OCEAN_STRESS_STAGGER controls where the stresses
that the ice offers up are discretized.  Many of the new options are not fully
implemented and result in error messages.  Also added comments describing many
of the elements in the ice_data_type, and reordered the optional arguments to
ice_model_restart to agree with the usage in the Tikal release of the shared
code. The answers are bitwise identical with the default settings.