Changes to accommodate C-grid stresses.
  Added code to handle the input of C-grid wind stresses to SIS2, for either
B-grid or C-grid discretizations of the dynamcics. Added a run-time parameter,
ICE_DEL_SH_MIN_SCALE, to permit the testing of various prescriptions for the
minimum shear rates used in the denominators of the rheology. Corrected the
positions in the register_restart_field calls for IST%u_ice_C and IST%v_ice_C.
All answers are bitwise identical in configurations that worked previously.