Use C-grid stresses internal with C-grid dynamics.
  Use C-grid discretizations of wind stresses in the code instead of sometimes
interpolating first to a B-grid.  When the B-grid stresses are reported back to
the ocean, this leads to order of arithmetic changes.  Also, made the
diagnostics FA_X and FA_Y apply to either B-grid or C-grid wind stresses.  The
variables flux_u_top_bgrid and part_size_uv are no longer used with C-grid
dynamics.  All external interfaces are unchanged. Answers with C-grid dynamics
are mathematically identical, but differ bitwise.