Use C-grid velocities directly in ice_bergs.F90.
  Added a new optional argument to icebergs_run to specify the staggering of the
wind stress separately from the staggering of the ocean and ice velocities, and
call this routine from ice_model with this value set to the value that is being
reported by the ocean.  Additionally, when the sea ice is on a C-grid, it now
passes c-grid velocities to the icebergs code, instead of interpolating to
b-grid velocities first.  Answers are unchanged unless both icebergs and the
c-grid discretization of the ice dynamics or a c-grid staggering of the
ice-ocean stresses are used.