Made several SIS2 restart fields non-mandatory.
  Made a number of variables that are present in SIS2 restart files but were
absent from SIS1 into non-mandatory variables and initialized them sensibly so
that SIS2 can now be restart from a SIS1 restart file, and a run with a B-grid
spatial discretization can be started from a C-grid discretization and
vice-versa.  All external interfaces are unchanged and all answers are bitwise
identical in test cases.