Fixed extrapolate_metric for non-zero fill values.
  Added an optional argument to extrapolate_metric to specify the values that
are used to fill in the arrays before they are read.  For many grid lengths, 0
is an appropriate fill value, but for latitudes and longitudes were being set to
999 because 0 is a legitimate value; for those arrays, no appropriate values
were being set in halos that can not be updated with message passing.
Appropriate values are now being set in all cases.  This did not change the
overall model solution.  All external interfaces are unchanged and all answers
are bitwise identical.