Replaced t_ice with heat_ice where it is enthalpy.
  The input variable T_ice was being changed in place from a temperature to an
enthalpy (heat).  As a preliminary step toward changing the state variable T_ice
from ice temperature (which is not conserved) to ice enthalpy (which is
conserved), a new temporary array, heat_ice, was introduced where this
transformation occurs.  This does introduce a new 4-d array, so it should
eventually be removed to minimize the memory footprint.  All answers are bitwise
identical.