*Added rolling for small amounts of thick ice.
  Added code to cause the sea-ice to roll and thin when a single floe with all
of the mass in the category would be taller than it is wide and hence unstable.
This behavior is controlled by the new run-time variable SEA_ICE_ROLL_FACTOR,
that turns this on for a positive number and is off for 0.  The default, 1, is
the right value for the rolling of a single square ice flow.  In addition, the
expression for h_snow has been rewritten in a mathematically equivalent way that
is more robust to round-off level changes.  These modifications do change
answers for test cases using SIS2 with SIS1_TRANSPORT=.false., and new reference
solutions have been checked in.