+Split SIS2_ice_thm.F90 from ice_thm.F90.
  Split the SIS2_thermodynamics routines out of ice_thm.F90 and into a new file,
SIS2_ice_thm.F90, for greater transparency of which set of thermodynamics
routines are being called, and to facilitate the conversion from temperature to
enthalpy as the dominant state variable with SIS2_thermodynamics.  There is a
new external interface to set the parameters for SIS2_ice_thm, and some routines
have been moved into the new module SIS2_ice_thm.  All answers are bitwise
identical.