+Replaced IST%pen*IST%trn with IST%sw_abs_ocn.
  Revised the shortwave radiation code to make use of the absorption arrays in
place of the IST%pen and IST%trn, which were specific to models with only a few
ice layers.  These changes are mathematically equivalent, but cause changes in
the answers at the bit level.  Also replaced the little-used diagnostic sw_trn
with sw_abs_ocn and added the diagnostic sw_abs_sfc for the fractional shortwave
absorbed by the ocean and surface skin layer.  New reference solutions were
checked in.