+Addition of SIS_tracer_registry.
  Added the first check-in of SIS_tracer_registry.F90, along with the addition
of a pointer to a SIS_tracxer_registry_type in the ice_state_type.  Also
promoted enth_snow from a 3-d array to a 4-d array with outer dimension of 1,
and added calls registring enth_snow and enth_ice as tracers, although the
registry does not cause anything to be done as yet.  All answers are bitwise
identical, but there are multiple interface changes.