+Use registry to pass tracers to ice_transport
  Use a SIS_tracer_registry_type argument to ice_transport, in place of heat_ice
and heat_snow, and added a new routine, get_SIS_tracer_pointer, as an interim
step to retrieve the tracers from the registry and use the current advection
code.  Also fixed the declaraion of the input target array index ranges in
register_SIS_tracer. All answers are bitwise identical, but there are multiple
interface changes.