+Added prognostic ice salinity to SIS2.
  Added a prognostic sea-ice salinity, along with a new run-time argument,
ICE_RELATIVE_SALINITY, that controls whether sea-ice takes a constant salinity
or whether its initial salinity is proportional to the salinity of the water
from which it formed.  With a constant salinity, there are order-of-operation
differences in the calculation of the salt flux to the ocean.  Answers change in
all test cases that use the SIS2 thermodynamics code, and new reference
solutions are checked in.