*+Pass integrated ice masses to dynamics.
  Replaced the average thickness arguments to ice_B_dynamics and ice_C_dynamics
with the integrated ice and snow masses per unit total (ice-covered plus
ice-free) area.  The resulting calculations are mathematically identical, but
differ in their order of arithmetic.  There are changes to the interfaces, and
this changes answers (at the level of cascading roundoff) to all test cases
using SIS2; new reference solutions have been checked in.