+Added rescaling of thickness units.
  Added more careful documentation of the units of various thickness variables
and multiplication by G%H_to_kg_m2 or its reciprocal, as appropriate, to allow
for the thickness units be rescaled.  Eventually this should allow for
H_TO_KG_M2 to be changed by powers of two and still give bitwise identical
answers, but this does not work yet.  As long as H_TO_KG_M2 is 1 (the default)
all answers are bitwise identical, although there are few modest interface
changes.