Only convert ice enthalpy to temps for diagnostics
  Added code to only calculate 4-d arrays of sea-ice temperature if they are
needed for diagnostic output or restarts.  Once read-only restart fields are
being used, the temperatures will only be calculated for restarts.  Also added
halo updates for enthalpy immediately after initialization (this may be
unnecessary, but it was there for temperature) and eliminated checksums on the
temperatures. All test case solutions are bitwise identical.