*Limit laytemp_SIS2 temperatures below freezing
  Added a line in laytemp_SIS2 so that the estimated ice and snow temperatures
are always at or below the freezing point for that particular layer.  This has
the potential to change answers in any case that uses the SIS2 thermodynamics,
but in the test cases there are only answer changes with the coupled SIS2 test
cases that start out with sea-ice, and new reference solutions
have been checked in for these cases.