+Added CP_BRINE as a runtime argument to SIS2
  Subdivided the heat capacity variables in the SIS2 thermodynamics to allow for
the brine inside of the sea-ice to have a different heat capacity from the water
from which the ice is made.  This will allow the water heat capacity to be
changed without greatly complicating the numerical algorithms.  Also, corrected
the heat capacities used at various places in the SIS2 thermodynamics code to
use the heat capacity of water, not that of ice, although for now these heat
capacies are the same by default. By default, all answers are bitwise identical,
but new SIS_parameter_doc.all files were checked in.