*Modified PPM SIS_continuity to use total ice.
  Changed the algorithm for the directionally split versions of the
SIS_continuity solver to base the mass fluxes on the total ice masses summed
across thickness categories, with the relative contributions of each category
simply proportional to the upwind mass in that category.  This will change
answers, but not in any test cases that use the UPWIND_2D continuity scheme, but
since all existing test cases use that scheme, there are no updates to the
reference solutions.