Ice pressure code earrangements in ice_C_dynamics
  Made a set of bitwise identical code rearrangements in side of ice_C_dynamics
to separate out the limiting of the stress components into a new internally
called subroutine, limit_stresses. Also recast the variable pres_ice into the
product of pres_mice and mice; this will later optionally allow the ice strength
to evolve more gracefully over the course of an ice dynamics step due to the
evolution of the ice concntration.  All external interfaces are unchanged and
answers are bitwise identical.