*Modified SIS2 Cgrid input parameter defaults.
  Changed the default values of ICE_TDAMP_ELASTIC, PROJECT_ICE_DRAG_VEL, and
PROJECT_ICE_CONCENTRATION to the desired values.  This changes the actual value
used for PROJECT_ICE_CONCENTRATION in all C-grid test cases and for
ICE_TDAMP_ELASTIC in the Baltic test case.  Also changed the input value of
ICE_TDAMP_ELASTIC to the default for the OM4_025 test case.  New reference
solutions and SIS_parameter_doc files were checked in for all test cases that
use a C-grid discretization of the sea-ice.