+Added code to accommodate Cp_ice /= Cp_brine
  Added code to correctly calculate the sea-ice thermodynamics when the heat
capacity of ice and of brine are set do different values.  This situation had
previously led to fatal error messages.  In addition, the variable tfi was
changed to T_fr for consistency throughout the code, and the existing function
dTemp_dEnth was renamed dTemp_dEnth_EnS and a similar function with different
arguments, dTemp_dEnth_TS, was added.  The answers in all existing test cases are bitwise
identical.