*Relinearized ice heat capacity above freezing
  Redid the linearization of the enthalpy equation for initial ice temperatures
that are above freezing to use the value of the heat capacity just below the
freezing point rather than the heat capacity of pure ice.  This will change
answers slightly, but should help stabilize the system and make it more
continuous for small changes in temperatuere.  The old answers can be recovered
by setting the runtime parameter OLD_ICE_HEAT_CAPACITY to true.  New reference
solutions have been checked for all of the test cases that use SIS2
thermodynamics.