+*Enabled SIS2 to use the MOM6 equation-of-state code
  Added a new EOS_type element to the ice_thermo_type, and new arguments to
get_SIS2_themo_coefs to enable SIS2 to use the same equation of state code as
MOM6.  The EOS_type itself is initialized from SIS2 and not copied from MOM6 to
permit the ice and ocean to run on different PE-lists, but by comparing the
logged parameters it is easy to confirm that they are the same.  This new
capability is excercised during the ice nudging, and the duplicate copy of
calculate_density_derivs_wright in ice_model.F90 has been eliminated.
  The ice nudging was refactored and partly corrected, with a new parameter,
NUDGE_SEA_ICE_STABILITY, although further work needs to be done here.  This
will change answers in test cases that use sea-ice nudging.  All of the test
cases using SIS2 are bitwise identical, although cases that use sea-ice nudging
will see answer changes.