+*Revised sea-ice concentration nudging.
  Revised the ice nudging to work more plausibly, including a revision of the
functional form of the nudging, the application of nudging when there is too
much ice, and revised the run-time nudging parameters, including the addition of
NUDGE_SEA_ICE_TOLERANCE, and the renaming of several variables.  Also added
comments describing a number of existing variables.  All answers in test cases
are bitwise identical, but results with NUDGE_SEA_ICE=True will be very
different.