*Re-revised sea-ice concentration nudging.
  Reworked the sea-ice concentration nudging yet again, this time to simply add
the nudging heat fluxes to the input frazil or the diffive heat flux at the
bottom of the ice provided from the ocean.  This simplifies the code, but it
should also be functionally similar.  All answers in the test cases are bitwise
identical, although answers will change in cases with NUDGE_SEA_ICE=True.