+Refactored ice_resize_SIS2.
  Changed the arguments to ice_resize_SIS2 to pass a column of thicknesses and
properties, and moved the call to rebalance_ice_layers outside of
ice_resize_SIS2.  This cleans up the interfaces between modules, and will
allow for the SIS2_ice_thm code to be used in an essentially stand-alone
column mode.  Also encapsulated all calculations in ice_resize_SIS2 that are
done purely for debugging inside of a logical flag, that might subsequently
be turned into an argument. All answers are bitwise identical.