*Changed thickness unit conversions to reciprocals
  Replaced several thickness unit conversions by division with mathematically
equivalent thickness conversions by multiplication by the reciprocal.  This
would ordinarily change answers at the level of roundoff, but not in the special
case where the number of ice layers and H_TO_KG_M2 are both exact powers of 2,
as is the case in all the SIS2 tests cases.  All answers in the test cases
happen to be bitwise identical.