+Added register_SIS_tracer_pair
  Added a new subroutine to register tracers in pairs of snow and ice tracers
with similar meanings.  This preserves the current practice of using distinct
tracers for the ice and snow, while also providing the information that these
are fundamentally similar quantities.  Also corrected the introductory comments
for SIS_tracer_advect.F90.  This change involves a new public interface, but
all solutions are bitwise identical.