+Merged ntr_snow & ntr_ice in SIS_tracer_registry
  Combined the lists of snow and ice tracers into a single list, where
corresponding tracers have the same index, but snow or ice tracers without a
counterpart are paired with 0-sized, nameless entries.  This will allow the snow
and ice tracers to be treated separately while preserving the information about
their relationships. This change involves changes to the entries of a publicly
visible type, but all solutions are bitwise identical.