*Reordered flux accumulation in SIS2_thermo
  Rearranged the order with which three different fluxes from the ice to the
ocean are accumulated in SIS2_thermodynamics, so that they are summed across
thickness category before they are summed between the melting and frazil ice
phases, and the net melt rate is calculated separately before being combined
with the liquid precipitation going to the ocean.  This is mathematically
equivalent to the previous solutions, but changes answers at the level of
roundoff.