+Changed from using G%CatIce to IG%CatIce
  Changed from using the sea_ice_grid_type to the ice_grid_type for access to
CatIce, in preparation for the future complete separation of these to types so
the former can be shared with the ocean.  In addition, eliminated the use of
NIMEM_ and similar macros, in favor of the SZI_(G), etc., macros for code
consistency. All answers are bitwise identical, although this check-in does
involve changes to many interfaces to pass in an ice_grid_type variable and the
elimination of several rarely used macros.