+Changed from G%H_to_kg_m2 to IG%H_to_kg_m2
  Changed from using the sea_ice_grid_type to the ice_grid_type for access to
H_to_kg_m2, kg_m2_to_H, and H_subroundoff, in preparation for the future
complete separation of these to types so the former can be shared with the
ocean. All answers are bitwise identical, although this check-in does involve
the elimination of 3 fields in the sea_ice_grid_type and the addition of a new
argument to several publicly visible subroutines.