+Passed an ice_grid_type to more subroutines
  Passed an ice_grid_type array into subroutines that had previously been using
it as a sub-element of the sea_ice_grid_type, except in cases where the
sea_ice_grid_type itself was being uses as a subelement of the ice_data type.
This is in preparation for the removal of the ice_grid_type as a subelement of
the sea_ice_grid_type, so that the latter can be made identical to its
counterpart in the MOM6 code. All answers are bitwise identical, although this
check-in does involve the addition of a new argument to several publicly visible
subroutines.