+Renamed Ice%mask to Ice%ocean_pt
  Renamed an element in the ice data type so that its meaning is much clearer -
it is a land-sea mask, and not an ice mask - and to verify that it was not being
used outside of SIS2.  Also, added comments highlighting where to change the
SIS2 code once a corrected version of flux_exchange.F90 has been made available
on the main development branch.  flux_exchange.F90 uses a module variable and an
element of the ice_data_type that would not otherwise exist, and should be
changed.  All answers are bitwise identical, although the name of an element
in a publicly visible type has been changed.