Combined initialize_SIS_masks & set_masked_metrics
  Rearranged the order of the initalization calls in initialize_fixed_SIS_grid
to enable the combination of initialize_SIS_masks and set_masked_metrics, in
another step toward combining the MOM6 and SIS2 initialization.  In addition,
initialize_SIS_masks was moved to the end of the file to be in the corresponding
position to its counterpart in MOM6.  All answers are bitwise identical.