Made GLOBAL_INDEXING a layoutParam & log G_EARTH
  GLOBAL_INDEXING should not change answers (although for now it does in MOM6),
so this run-time parameter should be a layout parameter.  Also added a call
logging the value of G_EARTH, which is currently pulled from the FMS constants
module but which will soon become an actual parameter. These modifications
change the logging in the MOM_parameter_doc files, but all answers are bitwise
identical.