*Corrected Celsius to Kelvin conversion
  SIS2 inhereted a conversion from Celsius to Kelvin that used 273.16, not
273.15, as the offset between Celsius and Kelvin. This oversight has now been
corrected, but as this changes the forcing, it changes solutions in all cases
using SIS2.  In addition, the code calculating the diagnostic of the average
lwdn was revamped for efficiency, and the use of constants_mod by SIS2 was
eliminated.  New reference solutions were added for all test cases using SIS2.