Explicitly initialize all of domore_u and domore_v
  Added code to explicitly initialize the entire arrays of domore_u and domore_v
to .false.   Curiously, the value used for the initialization of these logical
arrays does not matter to the solutions, because if .false. they are reevaluated
after the first halo update (and always on the first iteration, and if .true. a
number of fluxes are exactly 0 anyway.  Of the two choices, .false. is more
efficient in that it avoids extra calculations of 0 fluxes.  However, if they
are not explicitly initialized, this can be detected by performance analysis
software like valgrid, so this change is primarly to avoid generating warnings.
All answers are bitwise identical, and this should address both MOM6 issue #304
and MOM6 issue #305.