*Revised SIS version of set_grid_metrics_mercator
  Revised the SIS version of set_grid_metrics_mercator to use the same algorithm
as in the MOM version.  The new version is more robust, but this would change
answers if SIS had ever been used with a Mercator grid.  The test cases, however,
all give the same answers.  Also added grid-metrics_chksum in SIS_grid_initialize, which is identical to
its counterpart in MOM_grid_initialize.