+Split avg_top_quantities and post_flux_diagnostics
  Separated the subroutine avg_top_quantities into avg_top_quantities and
post_flux_diagnostics, so that they can be called at separate points in the
algorithm.  Also corrected the intent of a number of variables and added local
copies of runoff, calving, and runoff_hflx to the ice_state_type.  All solutions
are bitwise identical although some interfaces and elements of public types have
been changed.