+Swap order of icebergs_run & post_flux_diagnostics
  Created two new diagnostic arrays, calving_preberg and calving_hflx_preberg,
to that the order of the calls to icebergs_run and post_flux_diagnostics no
longer matters, and then moved post_flux_diagnostics later in SIS_slow_thermo.
All of this is in preparation for moving the call to icebergs_run outside of
SIS_slow_thermo.  All answers, and diagnostics are bitwise identical, although
there are modifications to a publicly visible data type.