+Created fast_radiation_diagnostics
  Made multiple interface changes, including the creation of a new subroutine,
fast_radiation_diagnostics, the elimination of the ice_data_type argument
from all of the routines in SIS_fast_thermo, the elimination of lwdn and swdn
from the fast_ice_avg_type.  All solutions are bitwise identical, but there are
numerous interface changes, and there may be slight differences in the LWDN and
SWDN diagnostics.