+Restored SW_DIR and SW_DIF diagnostics
  Added code to calculate the SW_DIR and SW_DIF diagnostics and restored them as
registered diagnostics.  Also moved frazil variables into the
ocean_sfc_state_type (for the value input from the ocean) and into the
fast_ice_avg_type (for the forcing that has not yet been appled, and out of the
catchall ice_state_type.  All answers are bitwise identical, but there are
changes to several publicly visible data structures and restoring the
diagnostics leads to changes in the available_diags files.