Set ocean albedo in set_ice_surface_state
  Added calls to set the ocean albedo in set_ice_surface_state and removed them
from ice_model_init.  Also wrapped the call to set_ocean_albedo in
set_fast_ocean_sfc_properties in a new logical flag that should be able to
determine when these updates are necessary.  For now this flag is hard-coded to
.true. because setting it dynamically changes answers unexpectedly. All answers
are bitwise identical and there are no interface changes apart from a new
element added to the ocean_state_type.