+Copy ice_transport_CSp into slow_thermo_CS
  Added a pointer to the ice_transport_CS to the slow_thermo_CS type, and added
a subroutine (SIS_slow_thermo_set_ptrs) to set this pointer and a new pointer to
a SIS_sum_output_CS in the slow_thermo_CS.  Also a function to query for the
pointer to the ice_transport_CS that is contained within the dyn_trans_CS.
Finally, the SIS_slow_thermo code now only refers to its own pointers instead
of using the ones from the ice_state_type.  All answers are bitwise identical,
but the contents of two widely used types have changed and two new interfaces
have been added.