+Created ice_rad_register_restarts
  Moved coszen_nextrad and 3 controlling parameters related to radiation into
the ice_rad_type.  This required alloc_ice_rad to upgrated to
ice_rad_register_restarts so that coszen could continue to be registered (and
read) as a restart file.  All answers are bitwise identical, although a
subroutine was renamed and the contents of two shared data types were modified.