+Extracted ice_data_type from ice_diagnostics_init
  Eliminated the use of the ice_data_type from ice_diagnostics_init and added an
ice_grid_type argument.  Also refactored the calculation of the little-used
diagnostic, CELL_AREA, of the grid area normalized by the area of the  surface
of a sphere. This is a move toward cleanly separating the fast and slow ice
phases from the top-level controls.  All answers are bitwise identical, but a
shared interface has been changed.