+Added ice_boundary_types.F90
  Added a new module, ice_boundary_types.F90, for the simple boundary types used
when communicating via the FMS coupler.  This module includes the three
[component]_ice_boundary_type structures, along with the publicly visible chksum
routines for each.  (All of these can still be accessed indirectly via
ice_model_mod.) Because everything in this module is relatively simple, it can
be compiled with no dependencies on other SIS or MOM code.  All answers are
bitwise identical, but the module in which these three types can be found
directly has changed.