Rearranged routines in SIS2_ice_thm.F90
  Rearranged SIS2_ice_thm.F90 to separate those routines that update the
thermodynamic state of the model (and use SIS2_ice_thm_CS as their control
structure) from the routines that do more fundamental thermodynamic calculations
but do not update the state.  This file could be split into two modules at a
well defined and labeled point in the middle, but for now they are being kept
together to promote inlining and computational efficiency.  No answers or
interfaces are changed.