+Moved Time form ice_state_type to SIS_fast_CS
  Moved the sea-ice time out of the ice_state type, and put it into the
SIS_fast_CS.  There is a second pointer to a Time in the SIS_slow_CS, which may
point back to SIS_fast_CS if everything is on the same PEs, or independently
tracks the time if they are on different PEs.  All answers are bitwise
identical, but some types and interfaces have changed.