+Added ice_diags_fast_init
  Rearranged the location of several post_data and register_SIS_diag_field calls
to separate out the fast and slow ice diagnostics, and created a new routine,
ice_diags_fast_init, to deal with the registration for the diagnostics that will
occur on the fast ice PEs.  All solutions and diagnostics are identical, but a
new subroutine has been added, and the order of entries in the
SIS.available_diags files have changed substantially.