+Added FAST_ICE_RESTARTFILE and split restart files
  Added the ability to write out the fast and slow sea-ice elements in separate
restart files, in preparation for having the fast and slow components run on
separate PEs.  Also changed the treatment of restart registration to only occur
if an associated restart_file_type is passed in to the various registration
routines.  All answers are bitwise identical, and the sea ice restart
capabilties have been tested to work with either separate or unified sea ice
restart files.  The default behavior is unchanged, but there is a new run-time
parameter, and the SIS_parameter_doc.all files have changed.