+Added simple_OSS_type
  Added a new type to contain the ocean data that is made available to the
atmosphere and the fast ice thermodynamics, plus the routines alloc_simple_OSS
and dealloc_simple_OSS to create and destroy this type, and copy_OSS_to_sOSS
to populate it from and ocean_sfc_state_type.  In addition, an new element,
Cgrid_dyn was added to the ocean_sfc_state_type.  All answers are bitwise
identical, but new types and interfaces were added.