+Added axes_set_name to set_SIS_axes_info
  Added an optional argument to specify the axes-set that is being set up.  By
default, this reproduces the old behavior, but it also allows for there to be a
second set of axes for the fast ice that will not have namespace conflicts with
the existing slow ice diagnostic axes.  All answers are bitwise identical, but
a new optional argument has been added to one routine.