+Added copy_FIA_to_FIA and move FIA into fCS & sCS
  Added a new routine to copy one fast_ice_avg_type into another, and removed
the fast_ice_avg_type that was in the top level of the ice_data_type. There
are now fast_ice_avg_type elements declared in Ice%sCS%FIA and Ice%fCS%FIA, and
these may or may not point to one another, depending on the settings of
SINGLE_ICE_STATE_TYPE.  All answers are bitwise identical, but a new subroutine
has been added and some publicly visible types changed.