*Correctly rescale wind-stresses for changing ice
  Corrected the rescaling of the wind stresses over ice-free and ice-covered
water to agree with the stress from the atmosphere when the fractional ice
coverage changes due to sea-ice transport.  The previous code had done this
correctly for thermally-driven ice coverage changes, but was not correct when
there are 2 or more ice dynamics steps per coupling step.  This changes answers
for all test cases where DT_ICE_DYNAMICS is shorter than the coupling timestep.
New reference solutions have been checked in.