+Modified the update_ice_model_slow_up arguments
  Modified the interface to update_ice_model_slow_up by making the arguments
optional and adding a new optional argument to indicate when the Verona or
earlier verision of the coupler is being used, and added two new routines,
unpack_ocean_ice_boundary and exchange_slow_to_fast_ice to handle the slow-ice
part of update_ice_model_slow_up and the exchange between the fast and slow
parts.  The interfaces code will work identically when called the same way as
the old code.  All answers are bitwise identical.