+Modified SIS2 fast ice MOM_domains_init call
  Modified the call to MOM_domains_init in ice_model_init for the fast ice when
the fast and slow ice do not share types, so that that fast type is always
symmetric and by default has no halos.  Also, added the suffix "_FAST" to the
parameters that set the fast domain properties, as reflected in the
SIS_parameter.layout files. This requires the changes to the MOM6
infrastructure code that were checked in with MOM6 commit
NOAA-GFDL/MOM6@fae5b7bb696912e84cf97989b3a50b34ee0c242e.