+Added SST_C array to ice_data_type
  Added a new array with the ocean surfact temperatures in degrees Celsius to
the publicly visible ice_data_type.  This field is available on the slow ice
PEs, and has been added for diagnostic convenience in the coupler stocks code,
and to allow Ice%t_surf to be allocated with the fast ice layout.  All answer
are bitwise identical, but coordinated changes are required to SIS1 and SIS2
if this new array is being used by the coupler.