*Corrected SIS2 for use with Verona coupler
  Corrected an error (an omitted subroutine call) in the SIS2 interface for use
with the Verona and earlier versions of the coupler. This bug was first
introduced with NOAA-GFDL/SIS2@5dcc85501ddb851ec8a977e258d8c2650f16509f, 10 days
ago. The answers are now back to the correct values and agree with both the
Verona version and the new developmental versions of the coupler.